"refactor : JpaTransaction 올바른 경로설정"
