+++ src/main/java/com/team5/catdogeats/admins/service/impl/AdminAuthenticationServiceImpl.java
@@ -4,7 +4,7 @@ import com.team5.catdogeats.admins.domain.Admins;
 import com.team5.catdogeats.admins.domain.dto.*;
 import com.team5.catdogeats.admins.repository.AdminRepository;
 import com.team5.catdogeats.admins.service.AdminAuthenticationService;
-import com.team5.catdogeats.global.config.JpaTransactional;
+import com.team5.catdogeats.global.annotation.JpaTransactional;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
