+++ src/main/java/com/team5/catdogeats/admins/service/impl/AdminPasswordResetServiceImpl.java
@@ -6,7 +6,7 @@ import com.team5.catdogeats.admins.repository.AdminRepository;
 import com.team5.catdogeats.admins.service.AdminPasswordResetService;
 import com.team5.catdogeats.admins.service.RedisVerificationCodeService;
 import com.team5.catdogeats.admins.util.AdminUtils;
-import com.team5.catdogeats.global.config.JpaTransactional;
+import com.team5.catdogeats.global.annotation.JpaTransactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
